% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpi.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{CPI}
\format{Data frame with columns
\describe{
\item{period}{Date Monthly}
\item{index}{numeric Index 1982-1984=100}
}}
\source{
\url{https://research.stlouisfed.org/fred2/series/CPIAUCSL/downloaddata}
}
\usage{
cpi
}
\description{
The US Consumer Price Index for All Urban Consumers: All Items
}
\examples{
data(cpi)
}
\keyword{datasets}

