% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdp.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP}
\format{Data frame with columns
\describe{
\item{period}{Date Quarterly}
\item{gdp.cur}{numeric Gross domestic product (current dollars)}
\item{gdp.chained.2009}{numeric GDP in chained 2009 dollars}
}}
\source{
\url{http://www.bea.gov/national/}
}
\usage{
gdp
}
\description{
Gross Domestic Product of the United States.
}
\examples{
data(gdp)
}
\keyword{datasets}

