% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{gdp_export_database}
\alias{gdp_export_database}
\title{Export the results from a harvest as a duckdb database file}
\usage{
gdp_export_database(harvest, destdir = NULL)
}
\arguments{
\item{harvest}{the results from running the gdp_harvest function}

\item{destdir}{the location to save the database file}
}
\value{
file path to the database file
}
\description{
Export the results from a harvest as a duckdb database file
}
