% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{gdp_export_tables}
\alias{gdp_export_tables}
\title{Persist data in a temporary local database and export files in CSV and parquet format}
\usage{
gdp_export_tables(harvest, destdir = NULL)
}
\arguments{
\item{harvest}{the result from running the gdp_harvest function}

\item{destdir}{the directory where the CSV and parquet files are to be created}
}
\value{
path to export directory
}
\description{
Persist data in a temporary local database and export files in CSV and parquet format
}
