% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{gdp_harvest}
\alias{gdp_harvest}
\title{Harvest data from GDP API given an organisation}
\usage{
gdp_harvest(organisation = "202100-3054")
}
\arguments{
\item{organisation}{the identifier for the organisation}
}
\value{
a list of tbls
}
\description{
Harvest data from GDP API given an organisation
}
