% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{gdp_upload_files}
\alias{gdp_upload_files}
\title{Mirror files from a local directory to a S3 target path}
\usage{
gdp_upload_files(sourcedir, s3_targetdir = "kthb/projects/gdp")
}
\arguments{
\item{sourcedir}{the location of the local directory}

\item{s3_targetdir}{the target spec for the S3 path, by default "kthb/projects/gdp"}
}
\value{
exit status code from system command
}
\description{
Mirror files from a local directory to a S3 target path
}
