% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_lite.R
\name{read_gdp_lite}
\alias{read_gdp_lite}
\title{Read from a file or url a JSON conforming to a GDP API "lite" response}
\usage{
read_gdp_lite(
  url,
  entity = c("finansieradeaktiviteter", "utlysningar", "ansokningar")
)
}
\arguments{
\item{url}{param a file path or url pointing to a GDP API "lite" resource}

\item{entity}{the resource type, one of "finansieradeaktiviteter", "utlysningar" or "ansokningar"}
}
\value{
a list with tabular parsed results
}
\description{
Read from a file or url a JSON conforming to a GDP API "lite" response
}
