% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unemp.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{Unemployment}
\format{Data frame with columns
\describe{
\item{period}{Date Monthly}
\item{value}{numeric Percentage of the labor force unemployed}
}}
\source{
\url{https://research.stlouisfed.org/fred2/series/UNRATE/#}
}
\usage{
unemp
}
\description{
The US unemployment rate.
}
\examples{
data(unemp)
}
\keyword{datasets}

