% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_lite.R
\name{write_gdp_lite}
\alias{write_gdp_lite}
\title{Export three GDP API "lite" resources as a duckdb database file}
\usage{
write_gdp_lite(
  destfile = file.path(tempdir(), "gdp-lite.db"),
  url_proposals,
  url_calls,
  url_fundings
)
}
\arguments{
\item{destfile}{file path to location of db, by default a file "gdp-lite.db"
in the temp dir}

\item{url_proposals}{the file or url for the location of the JSON for proposals}

\item{url_calls}{the file or url for the location of the JSON for calls}

\item{url_fundings}{the file or url for the location of the JSON for fundings}
}
\description{
Export three GDP API "lite" resources as a duckdb database file
}
